--- src/__tests__/test.js
@@ -1,5 +1,0 @@
-describe('test myModule', () => {
-  it('should return 42', () => {
-    expect(true).toEqual(true);
-  });
-});
