+++ src/index.js
@@ -5,6 +5,8 @@ import yargs from 'yargs/yargs';
 import { createToc } from './toc/createToc';
 
 const homeDir = process.cwd();
+console.log(process.argv);
+console.log(homeDir);
 yargs(hideBin(process.argv))
   .scriptName('nmrium')
   .command(
@@ -12,13 +14,18 @@ yargs(hideBin(process.argv))
     'build toc.json',
     () => {},
     (argv) => {
-      createToc({ homeDir, ...argv });
+      createToc(homeDir, { ...argv });
     },
   )
   .option('clean', {
     alias: 'c',
     type: 'boolean',
-    description: 'Remove all .json files',
+    description: 'Remove all .json files before creating toc',
+  })
+  .option('datadir', {
+    alias: 'd',
+    type: 'string',
+    description: 'Home directory containing the data',
   })
   .option('nostructure', {
     type: 'boolean',
